Moved hs_is_open into, more suiting, lghs_website.utils. It is still accessible through lghs_website.hs_is_open too.
