+++ lghs_website/__init__.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from flask import Flask, render_template
 
 app = Flask(__name__)
@@ -12,51 +11,12 @@ app.config.from_envvar('LGHS_WEBSITE_CONFIG', silent=True)
 # Source: http://flask.pocoo.org/docs/0.10/config/#configuring-from-files (at the bottom of the section)
 
 from .views import general
-from .utils import weekday_name
+from .utils import hs_is_open
 
 
 @app.errorhandler(404)
 def not_found(error):
 	return render_template('404.html'), 404
-
-
-def hs_is_open(time:datetime=None, opening_hours:dict=None) -> bool:
-	"""Indicate if the HS is opened at said time. Only relies on hours.
-	
-	Args:
-		time =
-			A date (as datetime.datetime object) you want to know if the HS is open or not.
-			Defaults to: datetime.now()
-			
-		opening_hours = 
-			eg: {'monday': (13, 18), 'friday': (5, 7)}  # Tuples are (opening_hour, closing_hour).
-			Defaults to: app.config['OPENING_HOURS']
-	
-	
-	If a day when the HS is actually open is provided:
-	
-	>>> hs_is_open(datetime(2016, 1, 6, 15, 0), {'wednesday': (13, 18)})
-	True
-	>>> hs_is_open(datetime(2016, 1, 6, 12, 30), {'wednesday': (13, 18)})
-	False
-	
-	If a day when the HS doesn't open at all is provided:
-	>>> hs_is_open(datetime.datetime(2016, 1, 7, 15, 30), {'wednesday': (13, 18)})
-	False
-	"""
-	
-	if time is None:
-		time = datetime.now()
-	
-	if opening_hours is None:
-		opening_hours = app.config['OPENING_HOURS']
-	
-	start_hour, end_hour = opening_hours.get(
-		weekday_name[time.weekday()],  # Retrieve opening hours for said day
-		(0, 0)  # will always invalidate if the HS doesn't open on said day
-	)
-	
-	return start_hour <= time.hour < end_hour
 
 
 # NOTE: Let's try to avoid renew the cache too often if we can.
